requirements.txt
fastapi
uvicorn
httpx
firebase-admin
pydantic
apscheduler
python-dotenv
openai
apache-airflow
